Coinflip Logic Server Side
Implemented logic for coinflip on server side. On button click of flip client side generates a seed in the same sha256 hashed manner as the server. That gets pass via socket.emit() and the server uses that seed + serverSeed + nonce (process.env.nonce) to determine the outcome of the coinflip.

Next is to relay it all to the clients and render on DOM.